backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==1.4.46
psycopg2-binary==2.9.10
ccxt==4.1.76
openai==1.3.6
python-jose==3.3.0
python-dotenv==1.0.0
alembic==1.13.1
apscheduler==3.10.4
pydantic==2.11.7
pydantic-settings==2.1.0